Extract assert_clipboard_text to ClipboardHelpers
